Auto-start rtl_433 when RTL-SDR dongle plugged in.
Systemd service to run rtl_433 and restart it when it crashes, with modified udev rules to call the service when the RTL-SDR device is plugged into a USB port. The udev rules call systemctl directly to start/stop the service, rather than tagging the event and binding the service to the RTL-SDR device unit in systemd.  This because of a udev/systemd bug, which means that systemd never notices when the dongle is unplugged, and the device unit remains loaded indefinitely.