Adapt COVID sensor to fetch latest scale graphic
The infection scale graphic now has a date in the URL, which corresponds to the last data update, not necessarily today.  This is handled by adding an attribute to the COVID rate sensor to fetch the update date (with corresponding changes to the Python script to output both in JSON).  The date and postcode are then used to build the URL for a camera sensor to provide the graphic.  Location-specific secrets are rationalised.